+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - build(ios): Bump sentry-cocoa to 6.1.4 #1308
 - fix: Handle auto session tracking start on iOS #1308
 - feat: Use beforeNavigate in routing instrumentation to match behavior on JS #1313
+- fix: React Navigation Instrumentation starts initial transaction before navigator mount #1315
 
 ## 2.2.0-beta.0